[ENH] Add functionalities to create space and add data
An enhance has been made to the Liquid Earth Application Programming Interface. Two new functionalities were added to enable the creation of space on the platform and to add data to the space. All usages of the old `PostData` class were replaced with the new `AddDataPostData` and `AddNewSpacePostData` classes to accommodate these changes, also tests were updated accordingly.