[ENH] Add LiquidEarth integration to 3D visualization tutorial
Enhanced 3D visualization tutorial by adding instructions on how to upload models to LiquidEarth, a new feature in GemPy v3. Improved error handling in the LiquidEarthAPI to better assist users when a space is not found. The option to create a new space when one does not exist was also added.